Fixed pet movements

+++ Source Main 5.2/source/w_PetAction.h
@@ -12,8 +12,8 @@ public:
     virtual bool Release(OBJECT* obj, CHARACTER* Owner) { return false; }
 
 public:
-    virtual bool Model(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender) { return false; }
-    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender) { return false; }
-    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender) { return false; }
-    virtual bool Sound(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender) { return false; }
+    virtual bool Model(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender) { return false; }
+    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender) { return false; }
+    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender) { return false; }
+    virtual bool Sound(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender) { return false; }
 };
