+++ Source Main 5.2/source/w_PetActionCollecter_Add.h
@@ -18,10 +18,10 @@ public:
     virtual bool Release(OBJECT* obj, CHARACTER* Owner);
 
 public:
-    virtual bool Model(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
-    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
-    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
-    virtual bool Sound(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
+    virtual bool Model(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
+    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
+    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
+    virtual bool Sound(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
 
     //test
     void FindZen(OBJECT* obj);
@@ -53,8 +53,8 @@ protected:
     DWORD m_dwRoundCountDelay;
     ActionState m_state;
 
-    float m_fRadWidthStand;
-    float m_fRadWidthGet;
+    double m_fRadWidthStand;
+    double m_fRadWidthGet;
 
     //test
 };
@@ -69,8 +69,8 @@ public:
     virtual bool Release(OBJECT* obj, CHARACTER* Owner);
 
 public:
-    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
-    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
+    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
+    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
 
 protected:
     PetActionCollecterSkeleton();
