+++ Source Main 5.2/source/w_PetActionStand.h
@@ -15,9 +15,9 @@ public:
     virtual bool Release(OBJECT* obj, CHARACTER* Owner);
 
 public:
-    virtual bool Model(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
-    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
-    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, DWORD tick, bool bForceRender);
+    virtual bool Model(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
+    virtual bool Move(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
+    virtual bool Effect(OBJECT* obj, CHARACTER* Owner, int targetKey, double tick, bool bForceRender);
 
 private:
     PetActionStand();
